mitra_crm v0.1.0 MitraCrm.Crm

A process for handling a CRM activity. Manages stakeholders and their associated data.

Functions

add_engagement(pid, stakeholder, engagement)

Add an %Engagement{} specfied by engagement to the stakeholder in CRM process pid.

returns {:ok, [%Stakeholders{}]}.

add_new_engagement(pid, stakeholder, engagement_name, type_name, due_date)

Adds an engagement with name engagement_name, due on due_date of the type specified by type_name the stakeholder at crm processes pid. type_name should be one of the type names returned by get_engagement_types.

returns `{:ok, [%Stakeholders{}]}.

add_stakeholder(pid, stakeholder_name_opts)

add_stakeholder_raw(pid, stakeholder)

Add a %Stakeholder{} struct specified by stakeholder to the CRM process at pid.

returns {:ok, [%Stakeholders{}]}

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.

defer_engagement(pid, stakeholder, engagement)

Defers the engagement of stakeholder by deferrment. Creates a new engement with the same information as the original but with the date pushed out. Marks the original engagement's state to deferred.

returns `{:ok, [%Stakeholders{}]}.

from_json(json)

from_json!(json)

get_engagement_types(pid)

Asks the Crm processes at pid for its stored engagement types.

returns {:ok, [%EngagementType{}]}

get_neglected_stakeholders(pid, past_threshold, future_threshold)

Gets [%Stakeholder{}]s who have not been contacted in past_threshold days or will be contacted within future_threshold days.

returns {:ok, [%Stakeholder{}]}.

get_stakeholder_by_uuid(pid, uuid)

get_stakeholders(pid)

Get the stakeholders from the CRM processes at pid.

returns `{:ok, [%Stakeholders{}]}

get_upcoming_dates(pid)

Gets [%Stakeholder{}]s who have not been contacted in past_threshold days or will be contacted within future_threshold days.

returns {:ok, [%Stakeholder{}]}.

handle_call(msg, from, state)

adds an engagement to the stakeholder

init(list)

Invoked when the server is started. start_link/3 or start/3 will block until it returns.

init_arg is the argument term (second argument) passed to start_link/3.

Returning {:ok, state} will cause start_link/3 to return {:ok, pid} and the process to enter its loop.

Returning {:ok, state, timeout} is similar to {:ok, state} except handle_info(:timeout, state) will be called after timeout milliseconds if no messages are received within the timeout.

Returning {:ok, state, :hibernate} is similar to {:ok, state} except the process is hibernated before entering the loop. See c:handle_call/3 for more information on hibernation.

Returning {:ok, state, {:continue, continue}} is similar to {:ok, state} except that immediately after entering the loop the c:handle_continue/2 callback will be invoked with the value continue as first argument.

Returning :ignore will cause start_link/3 to return :ignore and the process will exit normally without entering the loop or calling c:terminate/2. If used when part of a supervision tree the parent supervisor will not fail to start nor immediately try to restart the GenServer. The remainder of the supervision tree will be started and so the GenServer should not be required by other processes. It can be started later with Supervisor.restart_child/2 as the child specification is saved in the parent supervisor. The main use cases for this are:

The GenServer is disabled by configuration but might be enabled later.

An error occurred and it will be handled by a different mechanism than the Supervisor. Likely this approach involves calling Supervisor.restart_child/2 after a delay to attempt a restart.

Returning {:stop, reason} will cause start_link/3 to return {:error, reason} and the process to exit with reason reason without entering the loop or calling c:terminate/2.

Callback implementation for GenServer.init/1.

load_stakeholders_from_file(pid, filename)

loads stakeholders as JSON data from filename.

returns {:ok, stakeholders}.

load_stakeholders_from_file!(pid, filename)

Loads stakeholders as JSON data from filename. Returns {:ok,}.

load_state_file(pid, filename)

Teslls the crm process of pid to load state from filename.

returns {:ok, state}.

load_state_file!(pid, filename)

Teslls the crm process of pid to load state from filename.

returns {:ok}.

load_state_proc(pid)

persist_state(pid, filename)

Tells the Crm proc at pid to write its state to filename.

returns {:ok}.

persist_state_pid(pid)

put_engagement_type(pid, type, description)

Adds a new %EngagementType{} with the type name of type and description to the Crm processes at pid.

returns `{:ok, [%EngagementType{}]}.

start_link(id)

Starts up a Crm process specified by id, which should probably map to something like ther user's id.

returns {:ok, pid}

start_link(id, filename, atom)

to_json(state)

to_json!(state)

upadte_stakeholder(pid, new_data)