mitra_crm v0.1.0 MitraCrm.Stakeholder

This is the Stakeholder module. Handles data of individual stakeholder.

Functions

add_engagement_to_stakeholders_timeline(stakeholders, engagement)

Updates multiple stakeholders adding engagement to each.

Returns [stakeholder].

add_engagement_to_timeline(stakeholder, engagement)

Updates stakeholder with a new engagement.

Returns {:ok, stakeholder}.

exists?(stakeholders, stakeholder)

exists?(stakeholders, uuid, atom)

from_json(json)

Imports a stakeholder from json.

Returns {:ok, stakeholder}.

from_json!(json)

Imports a stakeholder from json.

Returns stakeholder.

get_all_stakeholder_dates(stakeholders)

Gets the important upcoming dates for stakeholders.

Returns %{engagements: [%Engagement{}], important_dates: [%StakeholderDate{}]}.

get_neglected_stakeholders(stakeholders, past_threshold, future_threashold)

Determines if any stakeholders in stakeholders haven't been contacted in a while.

Returns [%Stakeholder{}] if the stakeholder hasn't been engaged since past_threshold or will be engaged before future_threshold (in days).

get_stakeholder_by_name(stakeholders, stakeholder_name)

Returns the first %Stakeholder{} with name from an array of stakeholders.

Returns {:ok, stakeholder} or {:error, :not_found}

get_stakeholder_by_uuid(stakeholders, stakeholder_uuid)

get_upcoming_dates(stakeholder)

Gets upcomming dates for stakeholder.

Returns {:ok, [%StakeholderDate]}.

get_upcoming_dates!(stakeholder)

Gets upcomming dates for stakeholder.

Returns [%StakeholderDate].

neglected_stakeholder?(stakeholder, past_threshold, future_threashold)

Determines if any stakeholder hasn't been contacted in a while.

Returns true if the stakeholder hasn't been engaged since past_threshold or will be engaged before future_threshold (in days).

new(name_opts)

Creates a new stakeholder with the given_name and family_name.

Returns {:ok, stakeholder}.

new(given_name, family_name)

new_concern(stakeholder, name, description, importance)

Updates stakeholder with a new concern specified by the concerns name, a description of the concern, and the importance of the concern (Range[0..9]).

Returns {:ok, stakeholder}.

new_date(stakeholder, date, description, reminder, annual)

Updates stakeholder with a new StakeholderDate specified by:

Parameters

date: date of the event in ISO-8601 format string.

description: description of the date (e.g. "Pet's Birthday").

reminder: how far in advance (in integer days) a reminder should occur.

annual: a true/false parameter of whether the event is annual.

Returns {:ok, stakeholder}.

replace_stakeholder(stakeholders, stakeholder)

to_json(stakeholder)

Exports stakeholder to json.

Returns {:ok, json}.

to_json!(stakeholder)

Exports stakeholder to json.

Returns json.

update_attributes(stakeholder, attribute, value)

Updates stakeholder with a new attribute whoose value is value.

Returns {:ok, stakeholder}.

update_concerns(stakeholder, concern)

Updates stakeholder with a new concern.

Returns {:ok, stakeholder}.

update_contact(stakeholder, contact_title, value)

Updates stakeholder contact information, setting contact_title to value.

Returns {:ok, stakeholder}.

update_dates(stakeholder, date)

Updates stakeholder with a date.

Returns {:ok, stakeholder}.

update_name(stakeholder, name)

update_relationship(stakeholder, relationship)

Updates stakeholder with a new relationship.

Returns {:ok, stakeholder}.

update_relationship_metric(stakeholder, dimension, value, focus)

Updates stakeholder relationship dimension to the value and focus specified.

Parameters

dimension is one of [:credability, :empathy, :reliability, :stakeholderfocus].

value is in Range[0..9].

focus is whether or not you are focusing on building that relationship dimension.

Notes

:stakeholderfocus is a metric around how much you focus on the stakeholder's needs as opposed to your own.

Returns {:ok, stakeholder}.